# src/main/resources/api.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Gardener API
  license:
    name: UTCN
servers:
  - url: http://gardener.swagger.io/v1
paths:
...
  /specimens:
    get:
      summary: List of all specimen objects of the given user
      operationId: getSpecimens
      tags:
        - specimens
      parameters:
        - in: query
          name: userToken
          required: true
          schema:
            type: string
          description: The token of the user who makes the request
      responses:
        '200':
          description: All specimens of the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecimenDtos"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Save or update a specimen object. If no id is provided, we will create a new specimen. If an id is provided, we will update all the fields of the specimen, this includes updating to 'null' the fields which are not sent.
      operationId: postSpecimen
      tags:
        - specimens
      parameters:
        - in: query
          name: userToken
          required: true
          schema:
            type: string
          description: The token of the user who makes the request
      requestBody:
        description: A specimen object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpecimenDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecimenDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /specimens/{id}:
    get:
      summary: One specimen
      operationId: getSpecimen
      tags:
        - specimens
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: Uuid of the specimen
        - in: query
          name: userToken
          required: true
          schema:
            type: string
          description: The token of the user who makes the request
      responses:
        '200':
          description: The specimen with the matching id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecimenDto"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete the specimen with the give id
      operationId: deleteSpecimen
      tags:
        - specimens
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: Uuid of the specimen
        - in: query
          name: userToken
          required: true
          schema:
            type: string
          description: The token of the user who makes the request
      responses:
        '200':
          description: The specimen with the matching id was deleted.
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    PlantTypeDto:
      required: ['name', 'featuresConfiguration', 'growingConfiguration']
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        image:
          type: string
        featuresConfiguration:
          $ref: '#/components/schemas/FeaturesConfigurationDto'
        growingConfiguration:
          $ref: '#/components/schemas/GrowingConfigurationDto'
        userId:
          type: string
          format: uuid
...
    FeaturesConfigurationDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        stemType:
          $ref: '#/components/schemas/StemType'
        stemSpikes:
          $ref: '#/components/schemas/StemSpikes'
        leafDivisionOfBlade:
          $ref: '#/components/schemas/LeafDivisionOfBlade'
        leafArrangementOnTheStem:
          $ref: '#/components/schemas/LeafArrangementOnTheStem'
        leafSurface:
          $ref: '#/components/schemas/LeafSurface'
        leafTexture:
          $ref: '#/components/schemas/LeafTexture'
        leafVeinType:
          $ref: '#/components/schemas/LeafVeinType'
        flowerExistance:
          $ref: '#/components/schemas/FlowerExistance'
        flowerShape:
          $ref: '#/components/schemas/FlowerShape'
        flowerPetals:
          $ref: '#/components/schemas/FlowerPetals'
        userId:
          type: string
          format: uuid
...
    GrowingConfigurationDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        soilType:
          $ref: '#/components/schemas/SoilType'
        soilQuantity:
          type: integer
          format: int32
        soilHumiditySummer:
          $ref: '#/components/schemas/FiveDegreeScale'
        soilHumidityWinter:
          $ref: '#/components/schemas/FiveDegreeScale'
        airHumidity:
          $ref: '#/components/schemas/FiveDegreeScale'
        airTemperatureSummer:
          type: integer
          format: int32
        airTemperatureWinter:
          type: integer
          format: int32
        lightIntensitySummer:
          $ref: '#/components/schemas/FiveDegreeScale'
        lightIntensityWinter:
          $ref: '#/components/schemas/FiveDegreeScale'
        userId:
          type: string
          format: uuid
...
    SpecimenDto:
      required: ['name', 'userId']
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        plantType:
          $ref: '#/components/schemas/PlantTypeDto'
        growingConfiguration:
          $ref: '#/components/schemas/GrowingConfigurationDto'
        season:
          $ref: '#/components/schemas/Season'
        userId:
          type: string
          format: uuid
    SpecimenDtos:
      type: array
      items:
        $ref: "#/components/schemas/SpecimenDto"
    StemType:
      type: string
      enum: ['WOODY', 'HERBACEOUS', 'NOT_VISIBLE']
    StemSpikes:
      type: string
      enum: ['NO', 'WOODY', 'SOFT']
    LeafDivisionOfBlade:
      type: string
      enum: ['SIMPLE', 'WITH_LEAFLETS', 'PINNATE_COMPOUND', 'PALMATE_COMPOUND' ,'DOUBLE_COMPOUND']
    LeafArrangementOnTheStem:
      type: string
      enum: ['BASAL', 'WHORLED', 'ALTERNATE', 'OPPOSITE']
    LeafSurface:
      type: string
      enum: ['SMOOTH', 'HAIRY', 'WARTY', 'PLEATED', 'ROUGHENED_RIDGES', 'GROOVED']
    LeafTexture:
      type: string
      enum: ['LEATHERY', 'STICKY', 'FLESHY', 'GRASSY', 'NEEDLE', 'RUGOSE', 'SPIKE']
    LeafVeinType:
      type: string
      enum: ['ARCUATE', 'CROSS_VENULATE', 'DICHTOMOUS', 'LONGITUDINAL', 'PALMATE', 'PARALLEL', 'PINNATE', 'RETICULATE', 'ROTATE']
    FlowerExistance:
      type: string
      enum: ['YES', 'NO']
    FlowerShape:
      type: string
      enum: ['REGULAR', 'IREGULAR']
    FlowerPetals:
      type: string
      enum: ['UNITED', 'SEPARATED', 'FAN']
    SoilType:
      type: string
      enum: ['SUCCULENTS', 'ORCHID', 'CONIFERS', 'REGULAR']
    FiveDegreeScale:
      type: string
      enum: ['VERY_LOW', 'LOW', 'MEDIUM', 'HIGH', 'VERY_HIGH']
    Season:
      type: string
      enum: ['SPRING', 'SUMMER', 'AUTUMN', 'WINTER']
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string